requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
fastapi==0.92.0
fastapi-csrf-protect==0.2.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
motor==3.1.1
passlib==1.7.4
pydantic==1.10.5
PyJWT==2.6.0
pymongo==4.3.3
python-decouple==3.7
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
